Feat:랜딩페이지 UI 컴포넌트 모두 합침
