+++ src/components/Landing/CommonLandingDetail.jsx
@@ -1,13 +1,221 @@
 import React from 'react';
-import LandingSection from './LandingSection';
-import { Card1, Card2, CardImg1, CardImg2 } from './Cards';
-import IntroSection from './IntroSection';
-import PointBox from './PointBox';
-import Title from './Title';
-import SubTitle from './SubTitle';
-import StartButton from './StartButton';
+import styled from 'styled-components';
 import cardImg1 from '../../assets/images/cardImg1.png';
 import cardImg2 from '../../assets/images/cardImg2.png';
+import PrimaryButton from '../common/Button/PrimaryButton';
+
+//랜딩 레이아웃
+const Layout = styled.div`
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 20px;
+`
+
+//LandingSection
+const LandingSection = styled(Layout)`
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	margin-top: 124px;
+
+	@media (max-width: 1248px) {
+		padding: 0 24px;
+	}
+
+	@media (max-width: 768px) {
+		padding: 0 20px;
+		margin-top: 106px;
+		min-height: calc(100lvh + 65px);
+	}
+`;
+
+//Cards
+const CardBase = styled.div`
+	width: 100%;
+	height: 324px;
+	border-radius: 16px;
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	justify-content: space-around;
+	background-color: var(--surface);
+
+	@media (min-width: 769px) and (max-width: 1023px) {
+		height: 440px;
+		justify-content: center;
+	}
+
+	@media (max-width: 768px) {
+		height: 362px;
+	}
+`;
+
+const Card1 = styled(CardBase)`
+	margin-bottom: 30px;
+
+	@media (min-width: 769px) and (max-width: 1023px) {
+		height: 440px;
+		flex-direction: column;
+		justify-content: center;
+	}
+
+	@media (max-width: 768px) {
+		padding: 24px 24px 51px 24px;
+		flex-direction: column;
+		justify-content: flex-start;
+		height: 100%;
+		margin-bottom: 24px;
+		overflow: hidden;
+	}
+`;
+
+const Card2 = styled(CardBase)`
+	justify-content: flex-start;
+	margin-bottom: 48px;
+
+	@media (min-width: 769px) and (max-width: 1023px) {
+		justify-content: center;
+	}
+
+	@media (max-width: 768px) {
+		padding: 24px 24px 51px 24px;
+		flex-direction: column;
+		justify-content: flex-start;
+		height: 100%;
+		margin-bottom: 37px;
+		overflow: hidden;
+	}
+`;
+
+const CardImg1 = styled.div`
+	background-size: auto;
+	background-position: center;
+	background-repeat: no-repeat;
+	width: 664px;
+	height: 186px;
+	margin: -12px;
+
+	@media (max-width: 768px) {
+		width: 369px;
+		height: 104px;
+		margin: -7.14px -7.245px;
+	}
+`;
+
+const CardImg2 = styled(CardImg1)`
+	background-size: cover;
+	background-position: center;
+	background-repeat: no-repeat;
+	width: 720px;
+	height: 220px;
+
+	@media (max-width: 768px) {
+		display: flex;
+		align-items: center;
+		justify-content: center;
+		width: 100%;
+		height: 90px;
+	}
+`;
+
+//IntroSection
+const IntroSection = styled.div`
+	display: flex;
+	flex-direction: column;
+	align-items: flex-start;
+	position: relative;
+	margin-left: 60px;
+
+	@media (min-width: 1024px) {
+		width: 340px;
+	}
+
+	@media (min-width: 769px) and (max-width: 1023px) {
+		margin-top: 40px;
+	}
+
+	@media (max-width: 768px) {
+		margin-bottom: 50px;
+	}
+`;
+
+//PointBox
+const PointBox = styled.div`
+	padding: 6px 12px;
+	margin-bottom: 16px;
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	background-color: var(--purple-600);
+	border-radius: 50px;
+	font-size: 14px;
+	color: var(--white);
+	font-weight: var(--font-bold);
+	letter-spacing: -0.005em;
+
+	@media (max-width: 768px) {
+		padding: 4px 12px;
+		top: -44px;
+	}
+`;
+
+//Title
+const Title = styled.h2`
+	padding: 0px;
+	margin-bottom: 8px;
+	font-weight: var(--font-bold);
+	font-size: 24px;
+	line-height: 36px;
+	letter-spacing: -0.01em;
+	color: var(--gray-900);
+
+	@media (max-width: 768px) {
+		font-size: 18px;
+		line-height: 28px;
+	}
+`;
+
+//SubTitle
+const SubTitle = styled.h3`
+	font-weight: var(--font-regular);
+	font-size: 18px;
+	line-height: 28px;
+	letter-spacing: -0.01em;
+	color: var(--gray-500);
+
+	@media (max-width: 768px) {
+		font-size: 15px;
+		line-height: 22px;
+	}
+`;
+
+//StartButton
+const StartButton = styled(PrimaryButton)`
+	margin-bottom: 174px;
+	line-height: 2.8rem;
+	font-size: 1.8rem;
+	white-space: nowrap;
+
+	span {
+		display: flex;
+		justify-content: center;
+		width: 160px;
+	}
+
+	@media (min-width: 769px) and (max-width: 1023px) {
+		width: 100%;
+		margin-bottom: 24px;
+		display: flex;
+		justify-content: center;
+	}
+
+	@media (max-width: 768px) {
+		width: 100%;
+		display: flex;
+		justify-content: center;
+		margin: 24px 0px;
+	}
+`;
 
 const CommonLandingDetail = ({ 
 	className,
